% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klipfolio_acquire_dashboard.R
\name{klipfolio_dashboards}
\alias{klipfolio_dashboards}
\title{klipfolio_dashboard}
\usage{
klipfolio_dashboards(user, password)
}
\arguments{
\item{user}{user (Klipfolio account email)}

\item{password}{password (Klipfolio account password)}
}
\value{
List of dashboards obtainable by \code{user} with password: \code{password} with id, name and description of dashboards
}
\description{
Aquire dashboards from the Klipfolio API
}
\examples{
email <- c("klipr.package@gmail.com")
password <- c("changeme")

klipfolio_dashboards(email,password)

}
