% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klipfolio_acquire_datasources.R
\name{klipfolio_datasources}
\alias{klipfolio_datasources}
\title{klipfolio_datasources}
\usage{
klipfolio_datasources(user, password)
}
\arguments{
\item{user}{user (Klipfolio account email)}

\item{password}{password (Klipfolio account password)}
}
\value{
List of dashboards obtainable by \code{user} with password: \code{password} with id, name,description and refresh information of datasrouces
}
\description{
Aquire datasources from the Klipfolio API
}
\examples{
email <- c("klipr.package@gmail.com")
password <- c("changeme")

klipfolio_datasources(email,password)

}
