% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klipfolio_acquire_groups.R
\name{klipfolio_groups}
\alias{klipfolio_groups}
\title{Aquire groups from the Klipfolio API}
\usage{
klipfolio_groups(user, password)
}
\arguments{
\item{user}{user (Klipfolio account email)}

\item{password}{password (Klipfolio account password)}
}
\value{
List of groups obtainable by \code{user} with password: \code{password} with name and description
}
\description{
Aquire groups from the Klipfolio API
}
\examples{
email <- c("klipr.package@gmail.com")
password <- c("changeme")

klipfolio_groups(email,password)
}
