% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klipfolio_acquire_klips.R
\name{klipfolio_klips}
\alias{klipfolio_klips}
\title{klipfolio_klips}
\usage{
klipfolio_klips(user, password)
}
\arguments{
\item{user}{user (Klipfolio account email)}

\item{password}{password (Klipfolio account password)}
}
\value{
List of dashboards obtainable by \code{user} with password: \code{password} with id, name and description of klips
}
\description{
Aquire klips from the Klipfolio API
}
\examples{
email <- c("klipr.package@gmail.com")
password <- c("changeme")

klipfolio_klips(email,password)
}
