% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klipfolio_put.R
\name{klipfolio_put}
\alias{klipfolio_put}
\title{Put data into datasource}
\usage{
klipfolio_put(data, datasourceID, user, password)
}
\arguments{
\item{data}{a data object}

\item{datasourceID}{id of datasource}

\item{user}{Klipfolio account email}

\item{password}{Klipfolio account password}
}
\value{
API response
}
\description{
Put data into datasource
}
\examples{
dat <- head(mtcars)
datasource <- "abcd1234"

#' email <- "klipfolior@klipfolior.com"
password <- "secret_klipfolior"

}
