% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klipfolio_acquire_users.R
\name{klipfolio_users}
\alias{klipfolio_users}
\title{klipfolio_users}
\usage{
klipfolio_users(user, password)
}
\arguments{
\item{user}{user (Klipfolio account email)}

\item{password}{password (Klipfolio account password)}
}
\value{
List of users obtainable by \code{user} with password: \code{password} with name, email and last login
}
\description{
Aquire users from the Klipfolio API
}
\examples{
email <- c("klipr.package@gmail.com")
password <- c("changeme")

klipfolio_users(email,password)

}
